pyproject.toml
[tool.poetry]
name = "tgp"
version = "0.1.0"
description = "Python implementation of the Terminal Graphics Protocol used by kitty"

authors = ["Chris Watson <cawatson1993@gmail.com>"]

readme = "README.md"

repository = "https://github.com/watzon/tgp"
homepage = "https://github.com/watzon/tgp"

keywords = [
    "terminal", "graphics", "terminal graphics protocol", "kitty",
    "shell", "images", "pil"
]

license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
pillow = "^6.2.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
